Add routing and landing page to frontend app
Refactored App.tsx to use React Router and moved authentication logic to a new AppRoutes component. Added a LandingPage with associated SCSS module and set up routes for login, registration, and landing pages.
