+++ printlib-frontend/src/App.tsx
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
-import LoginForm from './components/forms/LoginForm';
-import RegisterForm from './components/forms/RegisterForm';
-import { login, register } from './api/auth';
 import './styles/theme.scss';
+import { BrowserRouter as Router } from 'react-router-dom';
+import AppRoutes from './AppRoutes';
 
 function App() {
   useEffect(() => {
@@ -13,48 +12,13 @@ function App() {
     document.body.classList.toggle('light-theme');
   };
 
-
-  const handleLogin = async (username: string, password: string) => {
-    try {
-      const result = await login(username, password);
-      console.log('Login success:', result);
-    } catch (error) {
-      console.error('Login failed:', error);
-    }
-  };
-
-  const handleRegister = async (
-    username: string,
-    email: string,
-    password: string
-  ) => {
-    try {
-      const result = await register(username, email, password);
-      console.log("Registration success:", result);
-      setShowRegister(false); // go back to login
-    } catch (error) {
-      console.error("Registration failed:", error);
-    }
-  };
-
-  const [showRegister, setShowRegister] = React.useState(false);
-
   return (
-    <div>
+    <Router>
       <button onClick={toggleTheme}>Toggle Theme</button>
-      {showRegister ? (
-        <RegisterForm 
-          onRegister={handleRegister}
-          onBack={() => setShowRegister(false)}
-        />
-      ) : (
-        <LoginForm
-          onLogin={handleLogin}
-          onRegister={() => setShowRegister(true)}
-        />
-      )}
-    </div>
+      <AppRoutes />
+    </Router>
   );
 }
+
 
 export default App;
