+++ printlib-frontend/src/pages/LandingPage.module.scss
@@ -1,0 +1,13 @@
+@import '../styles/variables';
+
+.landingPage {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  background-color: var(--color-background);
+  color: var(--color-text);
+  font-family: $font-family;
+  padding: $spacing-lg;
+}
